Move addToPlaylist to common
