+++ src/containers/add-to-playlist/store/sagas.ts
@@ -3,7 +3,7 @@ import { takeEvery, put } from 'redux-saga/effects'
 import { fetchSavedPlaylists } from 'common/store/account/reducer'
 import { requiresAccount } from 'utils/sagaHelpers'
 
-import * as actions from './actions'
+import * as actions from '../../../common/store/ui/add-to-playlist/actions'
 
 function* handlRequestOpen(action: ReturnType<typeof actions.requestOpen>) {
   yield put(fetchSavedPlaylists())
