+++ src/containers/menu/TrackMenu.tsx
@@ -24,9 +24,9 @@ import {
   undoRepostTrack,
   shareTrack
 } from 'common/store/social/tracks/actions'
+import { requestOpen as openAddToPlaylist } from 'common/store/ui/add-to-playlist/actions'
 import { requestOpen as openTikTokModal } from 'common/store/ui/share-sound-to-tiktok-modal/slice'
 import { ToastContext } from 'components/toast/ToastContext'
-import { requestOpen as openAddToPlaylist } from 'containers/add-to-playlist/store/actions'
 import { getCollectionId } from 'containers/collection-page/store/selectors'
 import * as embedModalActions from 'containers/embed-modal/store/actions'
 import { newCollectionMetadata } from 'schemas'
