+++ src/containers/nav/mobile/TopLevelPage.tsx
@@ -4,9 +4,9 @@ import cn from 'classnames'
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 
+import { getIsOpen as getIsAddToPlaylistPageOpen } from 'common/store/ui/add-to-playlist/selectors'
 import { getModalVisibility } from 'common/store/ui/modals/slice'
 import AddToPlaylistPage from 'containers/add-to-playlist/mobile/AddToPlaylist'
-import { getIsOpen as getIsAddToPlaylistPageOpen } from 'containers/add-to-playlist/store/selectors'
 import ConfirmAudioToWAudioPage from 'containers/confirm-audio-to-waudio/mobile/ConfirmAudioToWAudioPage'
 import EditPlaylistPage from 'containers/edit-playlist/mobile/EditPlaylistPage'
 import useScrollLock from 'hooks/useScrollLock'
